pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sasha"
version = "0.1.0"
description = "Semi-Analytic Shallow Water Framework for Coastal Monitoring"
authors = [{name = "Mohamed ali GHANNAMI", email = "mohamed-ali.ghannami.1@ulaval.ca"}]
readme = "Readme.md"
requires-python = ">=3.7"
dependencies = [
    "numpy",
    "scipy",
    "pydantic",
    "colorlog",
    "matplotlib",
    "pyyaml",
    "jax",
    "jaxlib",
    "numba",
    "tensorflow",
    "intersect",
    "tqdm",
    "colour-science"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]